+++ src/lokalise/base_client.ts
@@ -15,6 +15,7 @@ export class BaseClient {
 		authHeader: "x-api-token",
 		enableCompression: false,
 		requestTimeout: undefined,
+		silent: false,
 	};
 
 	/**
@@ -22,8 +23,14 @@ export class BaseClient {
 	 * @param params - Configuration parameters including API key and optional features.
 	 * @throws Error if the API key is not provided or is empty.
 	 */
-	constructor(params: ClientParams) {
-		const { apiKey } = params;
+	constructor({
+		apiKey,
+		enableCompression = false,
+		silent = false,
+		tokenType = "",
+		host,
+		requestTimeout,
+	}: ClientParams) {
 		if (typeof apiKey !== "string" || apiKey.trim().length === 0) {
 			throw new Error(
 				"Instantiation failed: A non-empty API key or JWT must be provided.",
@@ -31,9 +38,10 @@ export class BaseClient {
 		}
 
 		this.clientData.token = apiKey;
-		this.clientData.enableCompression = params.enableCompression ?? false;
-		this.clientData.tokenType = params.tokenType ?? "";
-		this.clientData.host = params.host;
-		this.clientData.requestTimeout = params.requestTimeout;
+		this.clientData.enableCompression = enableCompression;
+		this.clientData.silent = silent;
+		this.clientData.tokenType = tokenType;
+		this.clientData.host = host;
+		this.clientData.requestTimeout = requestTimeout;
 	}
 }